refactor: extract shared namespace validation to common/validation
Move the duplicated namespace validation logic from both
wal_impl.go and cluster_config.go into a shared
validation.ValidateNamespace() function. This ensures consistent
validation rules (including ".." path traversal rejection) at
both the WAL and config layers.
